% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghraw.R
\name{ghraw_gsa_standards}
\alias{ghraw_gsa_standards}
\title{Get GSA IT Standards Profile list}
\usage{
ghraw_gsa_standards()
}
\value{
tibble: id, name, category, status, description, deployment_type
}
\description{
Get GSA IT Standards Profile list
}
