% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghraw.R
\name{ghraw_philly_contracts}
\alias{ghraw_philly_contracts}
\title{Get Philadelphia professional services contract data}
\usage{
ghraw_philly_contracts(fiscal_year = 2020, quarter = 4)
}
\arguments{
\item{fiscal_year}{Fiscal year (e.g. 2020)}

\item{quarter}{Quarter (1-4)}
}
\value{
tibble of contract data (columns vary by quarter)
}
\description{
Get Philadelphia professional services contract data
}
